# Use span instead of string splitting in beatmap parsing