… CodeAnalysis/BannedSymbols.txt
… osu.Game/Beatmaps/Formats/LegacyStoryboardDecoder.cs
… osu.Game/Beatmaps/Formats/Parsing.cs
… osu.Game/Rulesets/Objects/HitObjectParser.cs
… osu.Game/Rulesets/Objects/Legacy/ConvertHitObjectParser.cs
… osu.Game/Skinning/LegacyManiaSkinDecoder.cs
… osu.Game/Skinning/LegacySkinDecoder.cs
… osu.Game/osu.Game.csproj