@huoyaoyuan 2023-04-22
It looks like InspectCode version is too low for C# 11, which is necessary.