@huoyaoyuan 2023-04-23
OK it's a well known pitfall - do not compare span and string with `==`.