@huoyaoyuan 2024-02-04
Glad to see .NET 8 being adopted.

The parsing can be simplified more with .NET 8 and C# 12 feature. I'll revisit this, probably this month.